requirements.txt
# Core
Django==3.2.14
django-environ==0.9.0
gunicorn==20.1.0

# Persistence storage
psycopg2-binary==2.9.3
django-seed==0.3.1

# RESTful API
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0
django-cors-headers==3.13.0
drf-yasg==1.21.5
whitenoise==6.3.0

# Code quality
flake8==4.0.1
flake8-isort==4.1.1
coverage==6.4.2
black==22.6.0
pylint-django==2.5.3
pre-commit==2.20.0

# Certificates
cryptography~=37.0.4
